Bug fixes and linting error fixes
