+++ chrome/extension/inject.js
@@ -1,3 +1,9 @@
+import _ from 'lodash';
+
+function jsNameToCssName(name) {
+  return name.replace(/([A-Z])/g, '-$1').toLowerCase();
+}
+
 function updateStyles() {
   // Handles live update from color picker
   chrome.runtime.onMessage.addListener(
@@ -129,12 +135,51 @@ function getNotifications() {
   }, 60000);
 }
 
+function onLoadComplete() {
+  // We make all the injected styles inline for easier
+  // manipulation and then remove the removable-initial-styles class from body
+  chrome.storage.local.get(storage => {
+    const { style, display, fontFamily } = storage;
+    // Browsers should batch all these DOM updates as they are all consecutive
+    // First insert inline-styles
+    if (style) {
+      _.forEach(style, updateObject => {
+        const { color, selector } = updateObject;
+        const property = jsNameToCssName(updateObject.property);
+        const rgba = `rgba(${color.r},${color.g},${color.b},${color.a})`;
+        const elementsArray = document.querySelectorAll(selector);
+        elementsArray.forEach(element => {
+          element.style[property] = rgba;
+        });
+      });
+    }
+    // Then we insert inline-display
+    if (display) {
+      _.forEach(display, updateObject => {
+        const visible = updateObject.visible ? 'block' : 'none';
+        const elementsArray = document.querySelectorAll(updateObject.selector);
+        elementsArray.forEach(element => {
+          element.style.display = visible;
+        });
+      });
+    }
+    // Finally insert inline-fontFamily
+    if (fontFamily) {
+      document.body.style.fontFamily = `${fontFamily} !important`;
+    }
+
+    // When all inline-styles are applied we can remove the class from the body
+    document.body.classList.remove('removable-initial-styles');
+  });
+}
+
 window.addEventListener('load', () => {
   updateStyles();
   updateDisplay();
   updateFontFamily();
   reset();
   reloadPage();
+  onLoadComplete();
 
   /**
    * Set a delay whilst we wait for current page to compute all DOM
